requirements.txt
selenium
webdriver-manager
PyYAML
